+++ ui/dashboard/Panel.tsx
@@ -16,10 +16,11 @@ export function Panel({ panel }: { panel: PanelInfo }) {
         <div className="panel-header vertical-align-center">
           <span className="panel-name">{panel.name}</span>
           <span className="flex-right">
-            {formatDistanceToNow(panel.resultMeta?.lastRun, {
-              addSuffix: true,
-              includeSeconds: true,
-            })}
+            {panel.resultMeta?.lastRun &&
+              formatDistanceToNow(panel.resultMeta?.lastRun, {
+                addSuffix: true,
+                includeSeconds: true,
+              })}
           </span>
         </div>
       </div>
