+++ ui/style.css
@@ -348,25 +348,38 @@ header > div {
 }
 
 .pages {
-  .current-page {
+  .page-name {
+    margin: 0 5px 0 0;
+    border-left: 0;
+    border-top: 0;
+    border-right: 0;
+    border-radius: 0;
+    box-shadow: none !important;
+  }
+
+  .page-name--current {
+    font-size: 1rem;
+    border-color: #9f04ff;
     margin-right: 8px;
     margin-left: 8px;
     padding: 8px;
     border: 1px solid #9f04ff;
     border-bottom: 0;
   }
+}
 
-  .current-page .page-name {
-    font-size: 1rem;
-    border-color: #9f04ff;
-  }
+.page-mode {
+    padding-right: 15px;
+    text-decoration: none;
+    color: black;
+}
 
-  .page-name.button {
-    margin: 0 5px;
-    border-left: 0;
-    border-top: 0;
-    border-right: 0;
-  }
+.page-mode:hover {
+    color: black;
+}
+
+.page-mode--on {
+  font-weight: bold;
 }
 
 .panel {
